# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: newsletter
    description: Newsletter subscription operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /newsletter/subscribe:
    post:
      operationId: subscribeNewsletter
      tags: [newsletter]
      summary: Subscribe to newsletter
      description: Saves an email address to the newsletter subscriber list
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscribeRequest"
      responses:
        "201":
          description: Successfully subscribed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscribeResponse"
        "409":
          description: Already subscribed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Invalid email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    SubscribeRequest:
      type: object
      properties:
        email:
          type: string
          format: email
      required:
        - email
    SubscribeResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error